Python Pi task commit

+++ python/tasks/radius.py
@@ -1,6 +1,6 @@
 import math     #import math module to use the power and pi functions 
 
-radius = int(input("Enter your radius: "))    #get radius from user
+radius = float(input("Enter your radius: "))    #get radius from user
 
 area = math.pow(radius,2) * math.pi         #calculate the area pi * r^2
